Tests pass.
Still not sure about approach though. Having to add a Var type to each
possible data type we wish to use seems cumbersome. I wonder if this can
be factored out? And also if I can come up with a good way of letting
existing types play nicely without change. I.e. we should just be able
to use the built-in list type.